[#35] refactor: 데이터 가공 로직 분리
Loader을 이용해 불러온 원본 데이터를 hook을 이용해 가공하도록 함
관심사 분리 및 코드 가독성 향상을 위해 적용
